requirements.txt
flask
pymongo
python-dotenv
Requests
selenium
webdriver_manager
beautifulsoup4
gunicorn
